% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_glm_output_table.R
\name{create_glm_output_table}
\alias{create_glm_output_table}
\title{Creating Tidy Output Table from GLM Object}
\usage{
create_glm_output_table(mod_glm, dat)
}
\arguments{
\item{mod_glm}{This is an object of class 'glm'.}

\item{dat}{This is the dataframe used in fitting the glm object.}
}
\description{
This function is designed to extract the key information from a GLM object and compile it into a tidy dataframe
for use in R Markdown reports. Information includes coefficient estimates, standard errors, odds ratios,
Z statistics, and p-values.
}
\examples{
mod_glm = glm(as.factor(readmit_status) ~ `LesionSize (cm)` + Age + Sex + BMI, data = dat_single, family = "binomial")
glm_output = create_glm_output_table(mod_glm, dat_single)
}
\keyword{Output}
\keyword{Tidy}
