% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic_table1.R
\name{statistic_table1}
\alias{statistic_table1}
\title{Adding Test Statistic to table1}
\usage{
statistic_table1(x, ...)
}
\arguments{
\item{x}{This input is not necessary if used as an added function in the extra.cols argument in table1.
The input will automatically be detected by the function.}
}
\description{
This function allows you to add a column calculating the test statistic to the table1 function. The calculation is based on
one-way ANOVA for continuous variables and chi-squared for categorical variables.
}
\examples{
table1::table1(~ . | group, data = dat, extra.cols = c(`Test Statistic`= statistic_table1))
}
\keyword{Statistic}
\keyword{Test}
